app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
google-cloud-storage==2.16.0 